p3

+++ src/p2/P2_main.java
@@ -14,9 +14,9 @@ public class P2_main {
 		System.out.println(intResult);
 		intResult = a + b / c; // 3 int
 		System.out.println(intResult);
-		doubleResult = c / b; // 1 int wird zu 1.0 double
+		doubleResult = (double) c / b; // 1 int wird zu 1.0 double
 		System.out.println(doubleResult);
-		doubleResult = x + a / b; // 1.5 double
+		doubleResult = x + (double) a / b; // 1.5 double
 		System.out.println(doubleResult);
 		doubleResult = x + y * b; // 13.0 double
 		System.out.println(doubleResult);
